function_pointers

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -7,7 +7,7 @@
  * @size: array size
  * @action: pointer function
  */
-void array_iterator(int *array, size_t size, void (*action)(int))
+void array_iterator(int *array, int size, void (*action)(int))
 {
 	if (action != NULL)
 	{
